Bug fixes
Fixed a typo.
Added explicit mean-centering to the first two examples.
Fixed a bug in displaying eigenvectors: each eigenvector should be a column rather than a row in the eig() result. Now the results for calling PCA on A and getting eigenvectors of A^TA are the same. The mean centering figure is changed because of that.